+++ android/library/src/main/java/com/liferay/mobile/screens/base/AbstractFactory.java
@@ -11,4 +11,6 @@ public interface AbstractFactory {
 	BasicCredentialsStorageSharedPreferences getBasicCredentialsStorageSharedPreferences();
 
 	CredentialsStorage getOAuthCredentialsStorageSharedPreferences();
+
+	CredentialsStorage getCookieCredentialsStorageSharedPreferences();
 }