+++ android/library/src/main/java/com/liferay/mobile/screens/base/FactoryCE.java
@@ -2,6 +2,7 @@ package com.liferay.mobile.screens.base;
 
 import com.liferay.mobile.screens.context.storage.CredentialsStorage;
 import com.liferay.mobile.screens.context.storage.sharedPreferences.BasicCredentialsStorageSharedPreferences;
+import com.liferay.mobile.screens.context.storage.sharedPreferences.CookieCredentialsStorageSharedPreferences;
 import com.liferay.mobile.screens.context.storage.sharedPreferences.OAuthCredentialsStorageSharedPreferences;
 
 /**
@@ -17,5 +18,10 @@ public class FactoryCE implements AbstractFactory {
 	@Override
 	public CredentialsStorage getOAuthCredentialsStorageSharedPreferences() {
 		return new OAuthCredentialsStorageSharedPreferences();
+	}
+
+	@Override
+	public CredentialsStorage getCookieCredentialsStorageSharedPreferences() {
+		return new CookieCredentialsStorageSharedPreferences();
 	}
 }