+++ ios/Framework/Core/Context/LiferayConnectorFactory.swift
@@ -114,6 +114,8 @@ public protocol LiferayConnectorFactory {
 
 	func createAssetLoadByClassPKConnector(_ className: String, classPK: Int64) -> AssetLoadByClassPKLiferayConnector?
 
+	func createAssetLoadByPortletItemNameConnector(portletItemName: String) -> AssetLoadByPortletItemNameLiferayConnector?
+
 	func createRatingLoadByEntryIdConnector(
 		entryId: Int64,
 		ratingsGroupCount: Int32) -> RatingLoadByEntryIdLiferayConnector?
@@ -381,6 +383,10 @@ open class Liferay62ConnectorFactory: NSObject, LiferayConnectorFactory {
 		return nil
 	}
 
+	public func createAssetLoadByPortletItemNameConnector(portletItemName: String) -> AssetLoadByPortletItemNameLiferayConnector? {
+		print("Unsupported connector in Liferay 6.2: AssetLoadByPortletItemNameLiferayConnector")
+		return nil
+	}
 
 	open func createRatingLoadByClassPKConnector(
 			_ classPK: Int64,
@@ -790,4 +796,8 @@ open class Liferay70ConnectorFactory: NSObject, LiferayConnectorFactory {
 		return Liferay70AssetLoadByClassPKConnector(className: className, classPK: classPK)
 	}
 
+	public func createAssetLoadByPortletItemNameConnector(portletItemName: String) -> AssetLoadByPortletItemNameLiferayConnector? {
+		return Liferay70AssetLoadByPortletItemNameConnector(portletItemName: portletItemName)
+	}
+	
 }